lambda/requirements.txt
PyYAML
requests
boto3
jinja2
pytz
